fix(Usager): MAj donnee

+++ src/app/core/utils/app-http-interceptor.ts
@@ -6,6 +6,7 @@ import { Router } from "@angular/router";
 import { LocalStorageService } from "./local-stoarge-service";
 import { GlobalName } from "./global-name";
 import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
+import { ConfigService } from "./config-service";
 
 
 @Injectable()
@@ -35,17 +36,20 @@ export class AppHttpInterceptor implements HttpInterceptor {
                 console.log("Error Event");
               } else {
                 console.log(
-                  `error status : ${error.status} ${JSON.stringify(error.error)}`
+                  `error status : ${error.status}`
                 );
                 switch (error.status) {
                   case 401:
-                    this.lsService.remove(GlobalName.tokenName)
-                    this.lsService.remove(GlobalName.refreshTokenName)
-                    this.lsService.remove(GlobalName.expireIn)
-                    this.lsService.remove(GlobalName.userName)
-                    this.lsService.remove(GlobalName.exercice)
-                    this.modalService.dismissAll()
-                    this.router.navigate(['//login'])
+                    if (error.url && error.url.includes(ConfigService.toApiUrl())) {
+                      this.lsService.remove(GlobalName.tokenName)
+                      this.lsService.remove(GlobalName.refreshTokenName)
+                      this.lsService.remove(GlobalName.expireIn)
+                      this.lsService.remove(GlobalName.userName)
+                      this.lsService.remove(GlobalName.exercice)
+                      this.modalService.dismissAll()
+                      this.router.navigate(['/log-usager'])
+                    }
+                 
                     break;
                   case 403:
                     break;
